+++ src/main/kotlin/com/swiftmako/jormanager/repositories/HostRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface HostRepository : JpaRepository<Host, Int>
+interface HostRepository : JpaRepository<Host, Long>
